requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.8
django-cleanup==5.2.0
django-cors-headers==3.10.0
django-environ==0.8.1
django-heroku==0.3.1
django-tinymce==3.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
flake8==4.0.1
gunicorn==20.1.0
mccabe==0.6.1
Pillow==8.4.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
python-http-client==3.3.3
pytz==2021.3
sendgrid==6.8.3
sqlparse==0.4.2
starkbank-ecdsa==1.1.1
whitenoise==5.3.0
